refactor(LottoView):메시지 하드코딩 부분 템플릿으로 변경
